fixup! fix(ivy): ensure that the  correct `document` is available
