+++ packages/core/src/core_render3_private_export.ts
@@ -245,7 +245,7 @@ export {
 
 export {
   setDocument as ɵsetDocument
-} from './render3/interfaces/renderer';
+} from './render3/interfaces/document';
 
 // we reexport these symbols just so that they are retained during the dead code elimination
 // performed by rollup while it's creating fesm files.
